created context for user
